stock-prediction-platform/ml/requirements.txt
# Core ML
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
scipy>=1.12.0

# Boosters
xgboost==2.0.3
lightgbm==4.3.0
catboost==1.2.5

# Explainability
shap==0.45.1

# Data access
psycopg2-binary==2.9.9
yfinance==0.2.38

# Utilities
structlog==24.2.0
tenacity==8.3.0

# Object storage
boto3>=1.34.0

# Testing
moto[s3]>=5.0.0
